feat(tui): wire TUI into main binary as `aletheia tui` subcommand
Add TUI crate to root workspace and expose it as an optional dependency
behind the `tui` feature flag (default ON). The TUI entry point is
extracted into lib.rs so both the standalone `aletheia-tui` binary and
the `aletheia tui` subcommand share the same code path.

Headless deployments can disable with `--no-default-features`.
