+++ tui/Cargo.toml
@@ -1,10 +1,15 @@
 [package]
 name = "aletheia-tui"
-version = "0.1.0"
-edition = "2024"
-rust-version = "1.88"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
 description = "Terminal dashboard for the Aletheia distributed cognition system"
-license = "AGPL-3.0-or-later"
+publish = false
+
+[[bin]]
+name = "aletheia-tui"
+path = "src/main.rs"
 
 [dependencies]
 # Core TUI — ratatui 0.30 defaults to crossterm 0.29
@@ -12,15 +17,15 @@ ratatui = "0.30"
 crossterm = { version = "0.29", features = ["event-stream"] }
 
 # Async runtime
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "time", "sync"] }
+tokio = { workspace = true }
 
 # HTTP + SSE
 reqwest = { version = "0.12", features = ["stream", "json", "rustls-tls", "cookies"], default-features = false }
 reqwest-eventsource = "0.6"
 
 # Serialization
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
+serde = { workspace = true }
+serde_json = { workspace = true }
 
 # Markdown rendering
 pulldown-cmark = "0.13"
@@ -37,8 +42,8 @@ toml = "1.0"
 dirs = "6"
 
 # Logging (file-based, not stdout)
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
 tracing-appender = "0.2"
 
 # Error handling
@@ -55,9 +60,22 @@ fuzzy-matcher = "0.3"
 # Cross-platform URL/file opening
 open = "5"
 
-[workspace]
+# Standalone lint config — TUI was developed outside the workspace.
+# Matches workspace lints but relaxes pedantic + dead_code until cleanup.
+[lints.rust]
+future_incompatible = { level = "warn", priority = -1 }
+nonstandard_style = { level = "warn", priority = -1 }
+unsafe_code = "deny"
+dead_code = "allow"
+elided_lifetimes_in_paths = "allow"
 
-[profile.release]
-strip = true
-lto = "thin"
-opt-level = "z"
+[lints.clippy]
+dbg_macro = "deny"
+todo = "deny"
+unimplemented = "deny"
+exit = "deny"
+await_holding_lock = "warn"
+similar_names = "allow"
+unnested_or_patterns = "allow"
+too_many_lines = "allow"
+struct_excessive_bools = "allow"
